Extracted the private functions and converted to public.
