+++ src/long-multiplication-command-line/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod breakdown;
 pub mod generate;
 pub mod long_multiplication;
